Postgresql DB Integration with Django Project 1
Here is a version of the django project where the sqlite db is replaced with a postgresql db.

Please note that the postgresql db was supported locally, and will need to have the proper credentials altered. This can be done in the settings.py file within the ctran_project_env_1 app folder within the django project (lines 80 - 83).

Please also note that psycopg2 is a db adapter for Python that is required for this. I used "pip install psycopg2" to add to the project environment.